6378950 ddi_intr_block_enable() doesn't enable the 2nd interrupt handle's handler correctly on x86

+++ usr/src/uts/sun4/io/px/px_intr.c
@@ -2,9 +2,8 @@
  * CDDL HEADER START
  *
  * The contents of this file are subject to the terms of the
- * Common Development and Distribution License, Version 1.0 only
- * (the "License").  You may not use this file except in compliance
- * with the License.
+ * Common Development and Distribution License (the "License").
+ * You may not use this file except in compliance with the License.
  *
  * You can obtain a copy of the license at usr/src/OPENSOLARIS.LICENSE
  * or http://www.opensolaris.org/os/licensing.
@@ -20,7 +19,7 @@
  * CDDL HEADER END
  */
 /*
- * Copyright 2005 Sun Microsystems, Inc.  All rights reserved.
+ * Copyright 2006 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
 
@@ -634,7 +633,7 @@ px_msix_ops(dev_info_t *dip, dev_info_t *rdip, ddi_intr_op_t intr_op,
 		 * based on Resource Management policies.
 		 */
 		if ((ret = px_msi_alloc(px_p, rdip, hdlp->ih_inum,
-		    hdlp->ih_scratch1, hdlp->ih_scratch2, &msi_num,
+		    hdlp->ih_scratch1, (int)hdlp->ih_scratch2, &msi_num,
 		    (int *)result)) != DDI_SUCCESS) {
 			DBG(DBG_INTROPS, dip, "px_msix_ops: MSI allocation "
 			    "failed, rdip 0x%p inum 0x%x count 0x%x\n",
